requirements.txt
Appium-Python-Client==2.9.0
pytest==7.4.4
selenium==4.9.0
